Delete the user's Pref-HomepageSettings attribute on Reset
If the user's Pref-HomepageSettings is there but not empty, it won't
respect any new changes made to the global homepage settings.

However, the fix is still not visible due to referring to session cache
checks. The preferences are checked in the session before the database,
so users will still not see homepage changes.